chore: Move memberServerImage into the server service
