+++ docs/data-sources/member_server_image.md
@@ -1,5 +1,5 @@
 ---
-subcategory: "Member Server Image"
+subcategory: "Server"
 ---
