+++ docs/data-sources/member_server_images.md
@@ -1,5 +1,5 @@
 ---
-subcategory: "Member Server Image"
+subcategory: "Server"
 ---
