+++ internal/service/server/member_server_images_data_source.go
@@ -1,4 +1,4 @@
-package memberserverimage
+package server
 
 import (
 	"fmt"
