+++ internal/service/server/member_server_images_data_source_test.go
@@ -1,4 +1,4 @@
-package memberserverimage_test
+package server_test
 
 import (
 	"testing"
@@ -8,49 +8,50 @@ import (
 	. "github.com/terraform-providers/terraform-provider-ncloud/internal/acctest"
 )
 
-func TestAccDataSourceNcloudMemberServerImageBasic(t *testing.T) {
+func TestAccDataSourceNcloudMemberServerImagesBasic(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck:                 func() { TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: ProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceNcloudMemberServerImageConfig,
+				Config: testAccDataSourceNcloudMemberServerImagesConfig,
 				// ignore check: may be empty created data
 				SkipFunc: func() (bool, error) {
 					return SkipNoResultsTest, nil
 				},
 				Check: resource.ComposeTestCheckFunc(
-					TestAccCheckDataSourceID("data.ncloud_member_server_image.test"),
+					TestAccCheckDataSourceID("data.ncloud_member_server_images.member_server_images"),
 				),
 			},
 		},
 	})
 }
 
-func TestAccDataSourceNcloudMemberServerImageFilter(t *testing.T) {
+func TestAccDataSourceNcloudMemberServerImagesFilter(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck:                 func() { TestAccPreCheck(t) },
 		ProtoV6ProviderFactories: ProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceNcloudMemberServerImageConfigFilter,
+				Config: testAccDataSourceNcloudMemberServerImagesConfigFilter,
+				// ignore check: may be empty created data
 				SkipFunc: func() (bool, error) {
 					return SkipNoResultsTest, nil
 				},
 				Check: resource.ComposeTestCheckFunc(
-					TestAccCheckDataSourceID("data.ncloud_member_server_image.test"),
+					TestAccCheckDataSourceID("data.ncloud_member_server_images.member_server_images"),
 				),
 			},
 		},
 	})
 }
 
-var testAccDataSourceNcloudMemberServerImageConfig = `
-data "ncloud_member_server_image" "test" {}
+var testAccDataSourceNcloudMemberServerImagesConfig = `
+data "ncloud_member_server_images" "member_server_images" {}
 `
 
-var testAccDataSourceNcloudMemberServerImageConfigFilter = `
-data "ncloud_member_server_image" "member_server_images" {
+var testAccDataSourceNcloudMemberServerImagesConfigFilter = `
+data "ncloud_member_server_images" "member_server_images" {
 	filter {
 		name   = "platform_type"
 		values = ["LNX64"]
